// src/theme/type.ts
import { TextStyle } from "react-native";

export type ThemeType = {
  backgroundColor: string;
  bgReverse: string;
  buttonBG: string;
  activeButtonColor: string;
  inactiveButtonColor: string;
  largeRegularText: TextStyle;
  largeBoldText: TextStyle;
  largeMediumText: TextStyle;
  titleRegularText: TextStyle;
  titleBoldText: TextStyle;
  titleMediumText: TextStyle;
  normalRegularText: TextStyle;
  normalRegularLightText: TextStyle;
  normalBoldText: TextStyle;
  normalMediumText: TextStyle;
  smallButtonText: TextStyle;
  largeButtonText: TextStyle;
};